refactor(172): extract shared test fixtures
