+++ app/src/test/java/org/cescfe/numpairs/feature/game/presentation/GameViewModelOperandSelectionTest.kt
@@ -3,6 +3,7 @@ package org.cescfe.numpairs.feature.game.presentation
 import org.cescfe.numpairs.domain.puzzle.model.OperandSlot
 import org.cescfe.numpairs.domain.puzzle.model.Operator
 import org.cescfe.numpairs.domain.puzzle.support.puzzleWithRepeatedSixes
+import org.cescfe.numpairs.feature.game.presentation.support.enterStripValue
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNull
