+++ app/src/test/java/org/cescfe/numpairs/feature/game/presentation/GameViewModelOperatorAndTileStateTest.kt
@@ -1,6 +1,7 @@
 package org.cescfe.numpairs.feature.game.presentation
 
 import org.cescfe.numpairs.domain.puzzle.model.Operator
+import org.cescfe.numpairs.feature.game.presentation.support.enterStripValue
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNull
